clean and working on direction fix then move to transpose fix
tomorrow contigging
